util: Allow path separators and spaces in filenames
The path separators are necessary if we want to be able to organize
comics into books or chapters or categories.

The spaces are nice to have. I assume they haven't been allowed because
they require escaping in a terminal but any modern terminal provides
assistance for that during tab completion. Disallowing spaces here means
they can never be used. Instead a method could be added to an
appropriate class to sanitize filenames and replace spaces by default if
that is actually desired.