feat(webtoon): added hench comics
https://www.webtoons.com/en/canvas/hench/list?title_no=857225

Desi ( Short for Despair) is a henchmen for the evil dark lord Ivel. It's always been her dream to become the most cruel, violent, and evil lackey one could ever hope for. The only problem is that they can't seem to do anything right! When even the hero takes pity on you, what is a henchman supposed to do? UPDATES EVERY OTHER THURSDAY